feat: seperate db from dao
